Onboarding Page
Added Onboarding page component (work in progress).

Added an images folder for the slideshow for the Onboarding component.

Modified App.tsx to import the page and added a route to the page.

The only way to get to the page currently is to add /onboarding in the URL.  I have not yet figured out how to re-direct a user or new user to the Onboarding page upon login.